empirical-runs/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.5.0
